Commission Calculator
This is a simple calculator with no GUI. It has a very specific use. I made it because, at my current occupation, I get paid a commission but we do not get the update of our commission until the day after earning it. So I made this to be able to calculate it myself without using a normal calculator since there are specifics to the earnings system.